(ushell-p4-mergedown)
#rb all

[CL 11255512 by Stefan Boberg in Dev-Core branch]

+++ Engine/Source/Runtime/Core/Private/HAL/MallocBinnedArena.cpp
@@ -434,7 +434,7 @@ struct FMallocBinnedArena::Private
 					NodePool->SetCanary(FPoolInfoSmall::ECanary::SmallUnassigned, true, false);
 					Table.BlockOfBlockAllocationBits.FreeBit(OutBlockOfBlocksIndex);
 
-					uint64 AllocSize = Allocator.SmallPoolTables[InPoolIndex].PagesPlatformForBlockOfBlocks * Allocator.ArenaParams.AllocationGranularity;
+					uint64 AllocSize = static_cast<uint64>(Allocator.SmallPoolTables[InPoolIndex].PagesPlatformForBlockOfBlocks) * Allocator.ArenaParams.AllocationGranularity;
 
 					if (!bWasExhaused)
 					{
